create dedicated archive and unarchive (moveToMailbox) methods
while removeFromMailbox called with inbox as a parameter could be used as an
archiving method it bahaves differently than the new archive method.

removeFromMailbox only puts emails into the archive as a last resort (when
otherwise that email would be in no mailboxes); archive() always puts it into
the archive mailbox (and removes it from the inbox). If removeFromMailbox were
to be called as a substitude for archive it would lead to a weird behaviour
where the users presses archive but alter doesn’t find the email in the archive box.